+++ components/button/buttonHelpers.tsx
@@ -4,7 +4,7 @@ import { clsx } from 'clsx';
 import isNonNullable from '../_util/isNonNullable';
 import { cloneElement, isFragment } from '../_util/reactNode';
 import { PresetColors } from '../theme/interface';
-import type { BaseButtonProps, LegacyButtonType } from './button';
+import type { BaseButtonProps, LegacyButtonType } from './Button';
 
 const rxTwoCNChar = /^[\u4E00-\u9FA5]{2}$/;
 export const isTwoCNChar = rxTwoCNChar.test.bind(rxTwoCNChar);
